refactor: update Bloque entity and related components to use Long for IDs, enhance DTO structure, and improve repository methods

+++ src/main/java/com/coagronet/tipoBloque/repositories/TipoBloqueRepository.java
@@ -11,17 +11,18 @@ import com.coagronet.tipoBloque.TipoBloque;
 @Repository
 public interface TipoBloqueRepository extends JpaRepository<TipoBloque, Integer> {
 
-    List<TipoBloque> findByEmpresaIdAndEstadoIdNotOrderByIdAsc(
-            Long empresaId,
-            Integer estadoId);
+        List<TipoBloque> findByEmpresaIdAndEstadoIdNotOrderByIdAsc(
+                        Long empresaId,
+                        Integer estadoId);
 
-    Optional<TipoBloque> findByIdAndEmpresaIdAndEstadoIdNot(
-            Integer id,
-            Long empresaId,
-            Integer estadoId);
+        Optional<TipoBloque> findByIdAndEmpresaIdAndEstadoIdNot(
+                        Integer id,
+                        Long empresaId,
+                        Integer estadoId);
 
-    boolean existsByIdAndEmpresaIdAndEstadoIdNot(Integer id,
-            Long empresaId,
-            Integer estadoId);
+        boolean existsByIdAndEmpresaIdAndEstadoIdNot(
+                        Long id,
+                        Long empresaId,
+                        Integer estadoId);
 
 }
